Requestクラスを修正

+++ sql/001-create-table-and-load-data.sql
@@ -4,7 +4,7 @@ CREATE TABLE players
     id INTEGER unsigned AUTO_INCREMENT,
     name VARCHAR(20) NOT NULL,
     position VARCHAR(15) NOT NULL,
-    uniform_number INTEGER(5) NOT NULL,
+    uniform_number int(5) NOT NULL,
     prefecture VARCHAR(10) NOT NULL,
     PRIMARY KEY (id)
 );
